refactor: 사용하지 않는 주석 및 코드 제거
