+++ src/main/java/com/example/delivery/controller/HomeController.java
@@ -1,6 +1,6 @@
 package com.example.delivery.controller;
 
-import com.example.delivery.security.UserDetailsImpl;
+import com.example.delivery.global.security.UserDetailsImpl;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -9,11 +9,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class HomeController {
 
-//    @GetMapping("/")
-//    public String home(Model model, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-//        model.addAttribute("email",userDetails.getUsername());
-//        return "home";
-//    }
 @GetMapping("/")
 public String home(Model model, @AuthenticationPrincipal UserDetailsImpl userDetails) {
     if (userDetails != null) {
