+++ src/main/java/com/example/delivery/controller/OrderController.java
@@ -1,10 +1,8 @@
 package com.example.delivery.controller;
 
-import com.example.delivery.domain.Menu;
-import com.example.delivery.domain.Store;
 import com.example.delivery.dto.OrderMenuDto;
 import com.example.delivery.dto.StoreDto;
-import com.example.delivery.security.UserDetailsImpl;
+import com.example.delivery.global.security.UserDetailsImpl;
 import com.example.delivery.service.order.OrderService;
 import com.example.delivery.service.store.StoreService;
 import lombok.RequiredArgsConstructor;
