--- src/main/java/com/example/delivery/dto/LikeDto.java
@@ -1,29 +1,0 @@
-//package com.example.delivery.dto;
-//
-//import lombok.*;
-//
-//public class LikeDto {
-//
-//    @Getter
-//    @Setter
-//    @AllArgsConstructor
-//    @NoArgsConstructor
-//    @Builder
-//    public static class LikeRequestDto {
-//
-//        private Long storeId;
-//
-//    }
-//
-//
-//    @Getter
-//    @Setter
-//    @AllArgsConstructor
-//    @NoArgsConstructor
-//    @Builder
-//    public static class LikeResponseDto {
-//
-//        private boolean isLiked;
-//
-//    }
-//}
