--- src/main/java/com/example/delivery/exception/CustomException.java
@@ -1,16 +1,0 @@
-package com.example.delivery.exception;
-
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException{
-
-    private ErrorCode errorCode;
-
-    public CustomException(ErrorCode errorCode) {
-
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-
-    }
-}
