--- src/main/java/com/example/delivery/exception/ErrorCode.java
@@ -1,51 +1,0 @@
-package com.example.delivery.exception;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-import static org.springframework.http.HttpStatus.*;
-
-@AllArgsConstructor
-@Getter
-public enum ErrorCode {
-
-    // 400 BAD_REQUEST: 잘못된 요청
-    INVALID_REFRESH_TOKEN(BAD_REQUEST, "리프레시 토큰이 유효하지 않습니다"),
-    MISMATCH_REFRESH_TOKEN(BAD_REQUEST, "리프레시 토큰의 유저 정보가 일치하지 않습니다"),
-    WRONG_USERNAME(BAD_REQUEST, "아이디가 잘못 입력되었습니다."),
-    WRONG_PASSWORD(UNAUTHORIZED, "비밀번호가 잘못 입력되었습니다."),
-    WRONG_CONFIRM_PASSWORD(UNAUTHORIZED, "비밀번호확인이 잘못 되었습니다."),
-    WRONG_MULTIPARTFILE(BAD_REQUEST, "Multipartfile에 문제가 있습니다"),
-    WRONG_DTO(BAD_REQUEST,"DTO를 다시 확인해주세요"),
-
-    // 401 UNAUTHORIZED: 인증되지 않은 사용자
-    INVALID_AUTH_TOKEN(UNAUTHORIZED, "권한 정보가 없는 토큰입니다"),
-    UNAUTHORIZED_MEMBER(UNAUTHORIZED, "존재하지 않는 회원입니다."),
-    UNAUTHORIZED_STORE(UNAUTHORIZED, "존재하지 않는 가게입니다."),
-    UNAUTHORIZED_MENU(UNAUTHORIZED, "존재하지 않는 메뉴입니다."),
-
-    // 404 NOT_FOUND: 잘못된 리소스 접근
-    REFRESH_TOKEN_NOT_FOUND(NOT_FOUND, "로그아웃 된 사용자입니다"),
-    MEMBER_NOT_FOUND(NOT_FOUND, "해당 회원 정보를 찾을 수 없습니다."),
-    STORE_NOT_FOUND(NOT_FOUND, "해당 가게 정보를 찾을 수 없습니다."),
-    MENU_NOT_FOUND(NOT_FOUND, "해당 메뉴 정보를 찾을 수 없습니다."),
-    USER_NOT_FOUND(NOT_FOUND, "해당 사용자를 찾을 수 없습니다."),
-
-
-    // 409 CONFLICT: 중복된 리소스 (요청이 현재 서버 상태와 충돌될 때)
-    DUPLICATE_EMAIL(CONFLICT, "이미 존재하는 이메일입니다."),
-    DUPLICATE_NICKNAME(CONFLICT, "이미 존재하는 닉네임입니다."),
-    DUPLICATE_STORE(CONFLICT, "이미 존재하는 가게이름 입니다."),
-    DUPLICATE_LIKE(CONFLICT, "이미 좋아요를 누르셨습니다."),
-
-    // 500 INTERNAL SERVER ERROR
-    SERVER_ERROR(INTERNAL_SERVER_ERROR, "내부 서버 에러입니다.");
-
-
-
-    private final HttpStatus httpStatus;
-    private final String message;
-
-}
