--- src/main/java/com/example/delivery/exception/ErrorResponse.java
@@ -1,20 +1,0 @@
-package com.example.delivery.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@AllArgsConstructor
-@Getter
-public class ErrorResponse {
-
-    private LocalDateTime timeStamp;
-    private int status;
-    private String error;
-    private String message;
-
-}
-
-
-
