--- src/main/java/com/example/delivery/exception/GlobalExceptionHandler.java
@@ -1,8 +1,0 @@
-package com.example.delivery.exception;
-
-import org.springframework.web.bind.annotation.ControllerAdvice;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-    // 테스트
-}
