+++ src/main/java/com/example/delivery/global/exception/CustomException.java
@@ -1,0 +1,16 @@
+package com.example.delivery.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException{
+
+    private ErrorCode errorCode;
+
+    public CustomException(ErrorCode errorCode) {
+
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+
+    }
+}
