+++ src/main/java/com/example/delivery/global/exception/ErrorResponse.java
@@ -1,0 +1,20 @@
+package com.example.delivery.global.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@AllArgsConstructor
+@Getter
+public class ErrorResponse {
+
+    private LocalDateTime timeStamp;
+    private int status;
+    private String error;
+    private String message;
+
+}
+
+
+
