+++ src/main/java/com/example/delivery/global/exception/GlobalExceptionHandler.java
@@ -1,0 +1,8 @@
+package com.example.delivery.global.exception;
+
+import org.springframework.web.bind.annotation.ControllerAdvice;
+
+@ControllerAdvice
+public class GlobalExceptionHandler {
+    // 테스트
+}
