+++ src/main/java/com/example/delivery/service/user/UserServiceImpl.java
@@ -5,8 +5,8 @@ import com.example.delivery.constant.Role;
 import com.example.delivery.domain.User;
 import com.example.delivery.dto.UserDto;
 import com.example.delivery.dto.UserRegisterDto;
-import com.example.delivery.exception.CustomException;
-import com.example.delivery.exception.ErrorCode;
+import com.example.delivery.global.exception.CustomException;
+import com.example.delivery.global.exception.ErrorCode;
 import com.example.delivery.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
