--- src/main/java/com/example/delivery/tool/email/EmailTool.java
@@ -1,40 +1,0 @@
-package com.example.delivery.tool.email;
-
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-
-@Slf4j
-@Component
-@Transactional
-@RequiredArgsConstructor
-public class EmailTool {
-
-    private final JavaMailSender javaMailSender;
-
-
-    public void sendEmail(String email, String title, String text) {
-        SimpleMailMessage emailForm = createEmailForm(email, title, text);
-        try {
-            javaMailSender.send(emailForm);
-            log.info("이메일 발송 성공");
-        } catch (Exception e) {
-            log.error("이메일 발송 오류");
-        }
-    }
-
-
-    // 발신할 이메일 데이터 세팅
-    private SimpleMailMessage createEmailForm(String toEmail, String title, String text) {
-        SimpleMailMessage message = new SimpleMailMessage();
-        message.setTo(toEmail);
-        message.setSubject(title);
-        message.setText(text);
-
-        return message;
-    }
-}
